Changed the way motor PWM signals are generated.
The mini driver firmware now dynamically changes the frequency of the motor PWM signal. When a low duty cycle (speed) is requested the frequency will be slowed to a minimum of 20Hz and when a high duty cycle (speed) is requested the frequency will be raised to a maximum of approximately 80Hz. This gives the motors more torque at lower speeds and so should allow them to turn at lower speeds and overcome the friction in the motor gearboxes.